Add About, Features, and Home components with animations and responsive design
- Implement About component with mission, vision, and how it works sections
- Create Features component showcasing key features with icons and descriptions
- Develop Home component with a welcoming message and call-to-action button
- Utilize Framer Motion for smooth animations and transitions
- Incorporate responsive design for various screen sizes
- Add decorative elements and background gradients for enhanced visuals
