+++ components/Body.tsx
@@ -3,66 +3,19 @@ import Link from "next/link";
 import Image from "next/image";
 import { motion } from "framer-motion";
 import {
-  Mountain,
-  Waves,
-  Landmark,
-  Palmtree,
-  Compass,
-  Users,
   LoaderCircle,
   ChevronLeft,
   ChevronRight,
   ArrowRight,
-  Sparkles,
-  Star,
   Globe,
   MapPin,
-  Calendar,
-  Shield,
-  Zap,
+  Star,
+  Calendar
 } from "lucide-react";
 import { useRouter } from "next/navigation";
-
-const features = [
-  {
-    icon: Zap,
-    label: "AI-Powered Planning",
-    description:
-      "Personalized itineraries crafted by AI based on your preferences",
-    gradient: "from-purple-500 to-pink-500",
-  },
-  {
-    icon: Shield,
-    label: "Budget Control",
-    description: "Clear cost estimates across different budget tiers",
-    gradient: "from-green-500 to-emerald-500",
-  },
-  {
-    icon: Globe,
-    label: "Local Insights",
-    description: "Authentic recommendations for food, culture, and hidden gems",
-    gradient: "from-blue-500 to-cyan-500",
-  },
-  {
-    icon: Users,
-    label: "Group Planning",
-    description: "Collaborate with friends and family in real-time",
-    gradient: "from-orange-500 to-red-500",
-  },
-  {
-    icon: MapPin,
-    label: "Interactive Maps",
-    description: "Visual journey planning with state-wise exploration",
-    gradient: "from-indigo-500 to-purple-500",
-  },
-  {
-    icon: Star,
-    label: "Community Recommendations",
-    description:
-      "Community-driven suggestions for off-the-beaten-path experiences",
-    gradient: "from-orange-500 to-purple-500",
-  },
-];
+import Home from './Home';
+import About from './About';
+import Features from './Features';
 
 const stats = [
   { number: "5", label: "Happy Travelers" },
@@ -310,411 +263,9 @@ const Body: React.FC<BodyProps> = ({ sectionRefs, sections }) => {
       {/* Content */}
       <div className="relative z-10 text-white pt-16">
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8 sm:py-12 lg:py-16 space-y-12 sm:space-y-16 lg:space-y-20 min-h-screen">
-          <motion.div
-            initial={{ opacity: 0, y: 30 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 1, ease: "easeOut" }}
-            className="text-center space-y-6 sm:space-y-8 pt-8 sm:pt-12"
-          >
-            <motion.div
-              initial={{ scale: 0 }}
-              animate={{ scale: 1 }}
-              transition={{ duration: 0.8, delay: 0.2 }}
-              className="flex items-center justify-center space-x-3 mb-6"
-            >
-              <div className="w-8 h-0.5 bg-gradient-to-r from-transparent to-purple-400"></div>
-              <div className="flex items-center space-x-2 px-4 py-2 bg-white/10 backdrop-blur-md rounded-full border border-white/20">
-                <Sparkles className="w-5 h-5 text-orange-500 animate-pulse" />
-                <span className="text-sm font-medium text-purple-400 tracking-wide">
-                  AI-POWERED TRAVEL PLANNING
-                </span>
-                <Sparkles className="w-5 h-5 text-orange-500 animate-pulse" />
-              </div>
-              <div className="w-8 h-0.5 bg-gradient-to-l from-transparent to-green-400"></div>
-            </motion.div>
-
-            <motion.h1
-              initial={{ opacity: 0, y: 20 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ duration: 1, delay: 0.4 }}
-              className="text-4xl sm:text-5xl md:text-6xl lg:text-7xl font-bold tracking-tight leading-tight"
-            >
-              Welcome to{" "}
-              <span className="relative">
-                <span className="text-transparent bg-clip-text bg-gradient-to-r from-orange-400 to-green-400 animate-gradient">
-                  Itinerarly
-                </span>
-                <motion.div
-                  initial={{ scaleX: 0 }}
-                  animate={{ scaleX: 1 }}
-                  transition={{ duration: 1, delay: 1 }}
-                  className="absolute -bottom-2 left-0 right-0 h-1 bg-gradient-to-r from-orange-400 to-purple-400 rounded-full"
-                />
-              </span>
-            </motion.h1>
-
-            <motion.div
-              initial={{ opacity: 0, y: 20 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ duration: 1, delay: 0.6 }}
-              className="text-lg sm:text-xl md:text-2xl leading-relaxed font-light max-w-4xl mx-auto text-gray-200"
-            >
-              Your intelligent and{" "}
-              <span className="relative text-green-400 font-semibold">
-                personalized
-                <motion.div
-                  initial={{ scaleX: 0 }}
-                  animate={{ scaleX: 1 }}
-                  transition={{ duration: 0.8, delay: 1.2 }}
-                  className="absolute -bottom-1 left-0 right-0 h-0.5 bg-orange-400 rounded-full"
-                />
-              </span>{" "}
-              travel companion for exploring the diverse landscapes and rich
-              cultural heritage of India.
-            </motion.div>
-
-            {/* Stats Section */}
-            {/* <motion.div
-              initial={{ opacity: 0, y: 30 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.8, delay: 0.8 }}
-              className="grid grid-cols-1 sm:grid-cols-3 gap-6 sm:gap-8 mt-12 max-w-4xl mx-auto"
-            >
-              {stats.map((stat, index) => (
-                <motion.div
-                  key={stat.label}
-                  initial={{ opacity: 0, scale: 0.8 }}
-                  animate={{ opacity: 1, scale: 1 }}
-                  transition={{ duration: 0.5, delay: 1 + index * 0.1 }}
-                  className="text-center p-4 bg-white/5 backdrop-blur-sm rounded-2xl border border-white/10 hover:bg-white/10 transition-all duration-300"
-                >
-                  <div className="text-2xl sm:text-3xl font-bold text-orange-400 mb-1">
-                    {stat.number}
-                  </div>
-                  <div className="text-sm text-gray-300">{stat.label}</div>
-                </motion.div>
-              ))}
-            </motion.div> */}
-
-            <motion.div
-              initial={{ opacity: 0, y: 20 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.8, delay: 1.2 }}
-              className="flex flex-col sm:flex-row items-center justify-center space-y-4 sm:space-y-0 sm:space-x-4 mt-8"
-            >
-              <motion.div
-              whileHover={{ scale: 1.05 }}
-              whileTap={{ scale: 0.95 }}
-              className="relative group"
-              >
-
-              <div className="absolute -inset-1 bg-gradient-to-r from-orange-500/30 via-emerald-500/30 to-indigo-500/30 rounded-2xl opacity-0 group-hover:opacity-60 transition-all duration-300"></div>
-              
-   
-              <div className="absolute -inset-0.5 bg-gradient-to-r from-orange-400/40 via-emerald-400/40 to-indigo-400/40 rounded-2xl animate-pulse opacity-20"></div>
-              
-              <Link
-              href="/start"
-              className="relative inline-flex items-center space-x-3 px-10 py-5 text-xl font-bold text-white bg-gradient-to-r from-orange-600 via-emerald-600 to-indigo-600 rounded-2xl border-2 border-white/30 hover:border-white/50 transition-all duration-300 hover:shadow-xl backdrop-blur-sm min-w-[320px] justify-center group overflow-hidden shadow-lg"
-              >
-            
-              <div className="absolute inset-0 bg-gradient-to-r from-transparent via-white/15 to-transparent translate-x-[-100%] group-hover:translate-x-[100%] transition-transform duration-800 ease-in-out"></div>
-              
-              
-              <div className="absolute inset-0 bg-gradient-to-r from-white/5 via-white/10 to-white/5 opacity-0 group-hover:opacity-100 transition-opacity duration-300"></div>
-              
-             
-              <div className="absolute top-3 right-4 w-1.5 h-1.5 bg-white/80 rounded-full opacity-0 group-hover:opacity-100 animate-ping"></div>
-              <div className="absolute bottom-3 left-4 w-1 h-1 bg-orange-200/80 rounded-full opacity-0 group-hover:opacity-100 animate-ping delay-200"></div>
-              
-              <Sparkles className="w-7 h-7 group-hover:rotate-12 transition-transform duration-300 relative z-10 drop-shadow-sm" />
-              <span className="relative z-10 tracking-wide drop-shadow-sm">Start Your Journey</span>
-              <ArrowRight className="w-6 h-6 group-hover:translate-x-1 transition-transform duration-300 relative z-10 drop-shadow-sm" />
-              </Link>
-              </motion.div>
-            </motion.div>
-          </motion.div>
-
-          {/* About Section */}
-          <motion.div
-            id="about"
-            initial={{ opacity: 0, y: 30 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.8, delay: 0.4 }}
-            className="space-y-8 sm:space-y-12 py-10 sm:py-16 relative"
-          >
-            {/* Decorative elements */}
-            <div className="absolute top-10 right-10 w-40 h-40 bg-orange-500/10 rounded-full blur-3xl -z-10"></div>
-            <div className="absolute bottom-10 left-10 w-40 h-40 bg-blue-500/10 rounded-full blur-3xl -z-10"></div>
-
-            <div className="text-center space-y-4">
-              <motion.h2
-                initial={{ opacity: 0, y: 20 }}
-                animate={{ opacity: 1, y: 0 }}
-                transition={{ duration: 0.8, delay: 0.6 }}
-                className="text-3xl sm:text-4xl md:text-5xl font-bold text-white"
-              >
-                About{" "}
-                <span className="text-transparent bg-clip-text bg-gradient-to-r from-orange-400 to-green-400">
-                  Itinerarly
-                </span>
-              </motion.h2>
-              <motion.p
-                initial={{ opacity: 0, y: 20 }}
-                animate={{ opacity: 1, y: 0 }}
-                transition={{ duration: 0.8, delay: 0.8 }}
-                className="text-lg text-gray-300 max-w-3xl mx-auto"
-              >
-                Your AI-powered companion for creating personalized travel
-                experiences in India
-              </motion.p>
-            </div>
-
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-8 items-center">
-              <motion.div
-                initial={{ opacity: 0, x: -30 }}
-                animate={{ opacity: 1, x: 0 }}
-                transition={{ duration: 0.8, delay: 0.8 }}
-                className="space-y-6"
-              >
-                <div className="bg-white/5 backdrop-blur-lg p-6 rounded-2xl border border-white/10">
-                  <h3 className="text-xl sm:text-2xl font-semibold text-orange-400 mb-3">
-                    Our Mission
-                  </h3>
-                  <p className="text-gray-200 leading-relaxed">
-                    Itinerarly was created to revolutionize how travelers
-                    explore the rich cultural tapestry of India. We combine
-                    artificial intelligence with local expertise to create
-                    journeys that are both authentic and personalized to your
-                    preferences.
-                  </p>
-                </div>
-
-                <div className="bg-white/5 backdrop-blur-lg p-6 rounded-2xl border border-white/10">
-                  <h3 className="text-xl sm:text-2xl font-semibold text-green-400 mb-3">
-                    How It Works
-                  </h3>
-                  <p className="text-gray-200 leading-relaxed">
-                    Our AI analyzes your travel preferences, budget constraints,
-                    and desired experiences to craft the perfect itinerary.
-                    Whether you're seeking cultural immersion, adventure,
-                    relaxation, or culinary exploration, our platform adapts to
-                    your unique travel style.
-                  </p>
-                </div>
-
-                <div className="bg-white/5 backdrop-blur-lg p-6 rounded-2xl border border-white/10">
-                  <h3 className="text-xl sm:text-2xl font-semibold text-purple-400 mb-3">
-                    Our Vision
-                  </h3>
-                  <p className="text-gray-200 leading-relaxed">
-                    We envision a world where travel planning is stress-free and
-                    where every journey to India becomes a transformative
-                    experience. We aim to showcase both iconic landmarks and
-                    hidden gems that make India an extraordinary destination.
-                  </p>
-                </div>
-              </motion.div>
-
-              <motion.div
-                initial={{ opacity: 0, x: 30 }}
-                animate={{ opacity: 1, x: 0 }}
-                transition={{ duration: 0.8, delay: 1 }}
-                className="relative"
-              >
-                <div className="absolute -inset-4 bg-gradient-to-r from-orange-500/20 via-green-500/20 to-purple-500/20 rounded-3xl blur-xl"></div>
-                <div className="relative bg-white/5 backdrop-blur-lg rounded-3xl p-6 border border-white/20 overflow-hidden shadow-2xl">
-                  <div className="grid grid-cols-2 gap-4">
-                    <div className="aspect-square rounded-2xl overflow-hidden">
-                      <Image
-                        src="https://images.unsplash.com/photo-1524492412937-b28074a5d7da"
-                        alt="Taj Mahal"
-                        width={300}
-                        height={300}
-                        className="w-full h-full object-cover transition-transform duration-700 hover:scale-110"
-                      />
-                    </div>
-                    <div className="aspect-square rounded-2xl overflow-hidden">
-                      <Image
-                        src="https://images.unsplash.com/photo-1566552881560-0be862a7c445"
-                        alt="Kerala Backwaters"
-                        width={300}
-                        height={300}
-                        className="w-full h-full object-cover transition-transform duration-700 hover:scale-110"
-                      />
-                    </div>
-                    <div className="aspect-square rounded-2xl overflow-hidden">
-                      <Image
-                        src="https://images.unsplash.com/photo-1477587458883-47145ed94245"
-                        alt="Himalayan Mountains"
-                        width={300}
-                        height={300}
-                        className="w-full h-full object-cover transition-transform duration-700 hover:scale-110"
-                      />
-                    </div>
-                    <div className="aspect-square rounded-2xl overflow-hidden">
-                      <Image
-                        src="https://images.unsplash.com/photo-1587474260584-136574528ed5"
-                        alt="Indian Food"
-                        width={300}
-                        height={300}
-                        className="w-full h-full object-cover transition-transform duration-700 hover:scale-110"
-                      />
-                    </div>
-                  </div>
-
-                  <div className="mt-6 space-y-3">
-                    <div className="flex items-center space-x-2">
-                      <div className="w-4 h-4 rounded-full bg-orange-400"></div>
-                      <p className="text-sm text-gray-200">
-                        Cultural Experiences
-                      </p>
-                    </div>
-                    <div className="flex items-center space-x-2">
-                      <div className="w-4 h-4 rounded-full bg-green-400"></div>
-                      <p className="text-sm text-gray-200">Natural Wonders</p>
-                    </div>
-                    <div className="flex items-center space-x-2">
-                      <div className="w-4 h-4 rounded-full bg-purple-400"></div>
-                      <p className="text-sm text-gray-200">Culinary Journeys</p>
-                    </div>
-                    <div className="flex items-center space-x-2">
-                      <div className="w-4 h-4 rounded-full bg-blue-400"></div>
-                      <p className="text-sm text-gray-200">Adventure Travel</p>
-                    </div>
-                  </div>
-                </div>
-              </motion.div>
-            </div>
-          </motion.div>
-
-          {/* Enhanced Features Section - with proper ID */}
-          <motion.div
-            id="features"
-            initial={{ opacity: 0, y: 30 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.8, delay: 0.4 }}
-            className="space-y-8 sm:space-y-12 py-10 sm:py-16 relative"
-          >
-            {/* Decorative elements */}
-            <div className="absolute top-0 right-0 w-40 h-40 bg-green-500/10 rounded-full blur-3xl -z-10"></div>
-            <div className="absolute bottom-0 left-0 w-40 h-40 bg-purple-500/10 rounded-full blur-3xl -z-10"></div>
-
-            <div className="text-center space-y-4">
-              <motion.h2
-                initial={{ opacity: 0, y: 20 }}
-                animate={{ opacity: 1, y: 0 }}
-                transition={{ duration: 0.8, delay: 0.6 }}
-                className="text-3xl sm:text-4xl md:text-5xl font-bold text-white"
-              >
-                Why Choose{" "}
-                <span className="text-transparent bg-clip-text bg-gradient-to-r from-green-400 to-purple-400">
-                  Itinerarly
-                </span>
-                ?
-              </motion.h2>
-              <motion.p
-                initial={{ opacity: 0, y: 20 }}
-                animate={{ opacity: 1, y: 0 }}
-                transition={{ duration: 0.8, delay: 0.8 }}
-                className="text-lg text-gray-300 max-w-3xl mx-auto"
-              >
-                Experience the future of travel planning with our cutting-edge
-                features designed to make your Indian adventure unforgettable
-              </motion.p>
-            </div>
-
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 sm:gap-8 lg:gap-10">
-              {features.map((feature, index) => (
-                <motion.div
-                  key={feature.label}
-                  initial={{ opacity: 0, y: 30 }}
-                  animate={{ opacity: 1, y: 0 }}
-                  transition={{ duration: 0.6, delay: 0.8 + index * 0.1 }}
-                  whileHover={{
-                    scale: 1.03,
-                    y: -5,
-                    transition: { duration: 0.2 },
-                  }}
-                  className="group relative bg-white/5 backdrop-blur-lg p-6 sm:p-8 rounded-3xl border border-white/10 hover:border-white/30 transition-all duration-500 hover:shadow-2xl hover:shadow-white/10 overflow-hidden"
-                >
-                  {/* Background glow effect */}
-                  <div
-                    className={`absolute inset-0 bg-gradient-to-br ${feature.gradient} opacity-0 group-hover:opacity-5 transition-opacity duration-500`}
-                  />
-
-                  <div className="relative z-10">
-                    <div className="flex items-center space-x-4 mb-4">
-                      <div
-                        className={`p-3 sm:p-4 bg-gradient-to-br ${feature.gradient} rounded-2xl shadow-lg group-hover:scale-110 transition-transform duration-300`}
-                      >
-                        <feature.icon className="w-6 h-6 sm:w-7 sm:h-7 text-white" />
-                      </div>
-                      <h3 className="text-lg sm:text-xl font-bold text-white group-hover:text-transparent group-hover:bg-clip-text group-hover:bg-gradient-to-r group-hover:from-white group-hover:to-gray-300 transition-all duration-300">
-                        {feature.label}
-                      </h3>
-                    </div>
-                    <p className="text-sm sm:text-base text-gray-300 leading-relaxed group-hover:text-gray-200 transition-colors duration-300">
-                      {feature.description}
-                    </p>
-                  </div>
-                </motion.div>
-              ))}
-            </div>
-
-            <motion.div
-              initial={{ opacity: 0, y: 20 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.8, delay: 1.4 }}
-              className="bg-white/5 backdrop-blur-md p-6 sm:p-8 rounded-3xl border border-white/10 max-w-4xl mx-auto mt-12"
-            >
-              <h3 className="text-xl sm:text-2xl font-semibold text-white text-center mb-6">
-                Coming Soon
-              </h3>
-
-              <div className="grid grid-cols-1 sm:grid-cols-2 gap-6">
-                <div className="flex items-start space-x-4">
-                  <div className="p-3 bg-gradient-to-br from-yellow-500 to-orange-500 rounded-xl shadow-lg">
-                    <Calendar className="w-5 h-5 text-white" />
-                  </div>
-                  <div>
-                    <h4 className="text-lg font-medium text-orange-400 mb-2">
-                      Seasonal Recommendations
-                    </h4>
-                    <p className="text-gray-300 text-sm">
-                      Intelligent suggestions based on the best times to visit
-                      different regions of India
-                    </p>
-                  </div>
-                </div>
-
-                <div className="flex items-start space-x-4">
-                  <div className="p-3 bg-gradient-to-br from-blue-500 to-cyan-500 rounded-xl shadow-lg">
-                    <Users className="w-5 h-5 text-white" />
-                  </div>
-                  <div>
-                    <h4 className="text-lg font-medium text-cyan-400 mb-2">
-                      Family-Friendly Filters
-                    </h4>
-                    <p className="text-gray-300 text-sm">
-                      Specially curated itineraries for travelers with children
-                      of different age groups
-                    </p>
-                  </div>
-                </div>
-              </div>
-
-              <div className="mt-8 text-center">
-                <Link
-                  href="/start"
-                  className="inline-flex items-center space-x-2 px-6 py-3 text-sm font-semibold text-white bg-gradient-to-r from-purple-600/80 to-blue-600/80 backdrop-blur-sm rounded-xl border border-white/20 hover:border-white/40 transition-all duration-300 hover:shadow-lg"
-                >
-                  <span>Explore All Features</span>
-                  <ArrowRight className="w-4 h-4" />
-                </Link>
-              </div>
-            </motion.div>
-          </motion.div>
+          <Home />
+          <About />
+          <Features />
 
           {/* Community Images Section */}
           <motion.div
@@ -809,7 +360,7 @@ const Body: React.FC<BodyProps> = ({ sectionRefs, sections }) => {
                         loading={idx === 0 ? "eager" : "lazy"}
                         priority={idx === 0}
                         placeholder="blur"
-                        blurDataURL="data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAYEBQYFBAYGBQYHBwYIChAKCgkJChQODwwQFxQYGBcUFhYaHSUfGhsjHBYWICwgIyYnKSopGR8tMC0oMCUoKSj/2wBDAQcHBwoIChMKChMoGhYaKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCj/wAARCAAIAAoDASIAAhEBAxEB/8QAFQABAQAAAAAAAAAAAAAAAAAAAAv/xAAhEAACAQMDBQAAAAAAAAAAAAABAgMABAUGIWGRkqGx0f/EABUBAQEAAAAAAAAAAAAAAAAAAAMF/8QAGhEAAgIDAAAAAAAAAAAAAAAAAAECEgMRkf/aAAwDAQACEQMRAD8AltJagyeH0AthI5xdrLcNM91BF5pX2HaH9bcfaSXWGaRmknyJckliyjqTzSlT54b6bk+h0R//2Q=="
+                        blurDataURL="data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAYEBQYFBAYGBQYHBwYIChAKCgkJChQODwwQFxQYGBcUFhYaHSUfGhsjHBYWICwgIyYnKSopGR8tMC0oMCUoKSj/2wBDAQcHBwoIChMKChMoGhYaKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCj/wAARCAAIAAoDASIAAhEBAxEB/8QAFQABAQAAAAAAAAAAAAAAAAAAAAv/xAAhEAACAQMDBQAAAAAAAAAAAAABAgMABAUGIWGRkqGx0f/EABUBAQEAAAAAAAAAAAAAAAAAAAMF/8QAGhEAAgIDAAAAAAAAAAAAAAAAAAECEgMRkf/aAAwDAQACEQMRAD8AltJagyeH0AthI5xdrLcNM91BF5pX2HaH9bcfaSXWGaRmknyJckliyjqTzSlT54b6bk+h0R//2Q=="
                         sizes="(max-width: 768px) 100vw, (max-width: 1200px) 80vw, 800px"
                       />
                       {/* Image overlay with info */}
